extracted the TodoContainer component
