+++ src/App.css
@@ -42,46 +42,3 @@
   align-items: center;
   flex-direction: column;
 }
-
-.todoContainer {
-  width: 100%;
-  height: 100%;
-  overflow-y: auto;
-  background-color: #f8f9ff;
-  padding: 0 150px;
-  padding-top: 65px;
-}
-
-.initialButtonsContainer {
-  width: 100%;
-  font-family: Roboto, sans-serif;
-  display: inline-flex;
-  height: 36px;
-  justify-content: space-between;
-  margin-bottom: 34px;
-}
-
-.filterButtonsContainer {
-  display: inline-flex;
-  gap: 19px;
-}
-
-.todoList {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 33px;
-  align-items: stretch;
-  align-content: center;
-}
-
-.lowerOpacity {
-  opacity: 0.5;
-  pointer-events: none;
-}
-
-.todosDisplayController {
-  width: 100%;
-  padding: auto;
-  display: flex;
-  justify-content: center;
-}
